[nms] Update @fbcnms/alarms to 0.2.3
