+++ nms/app/packages/magmalte/app/state/AlarmsApiUtil.js
@@ -73,6 +73,7 @@ export const MagmaAlarmsApiUtil: ApiUtil = {
   createReceiver: async ({networkId, receiver}) => {
     await MagmaV1API.postNetworksByNetworkIdPrometheusAlertReceiver({
       networkId: nullthrows(networkId),
+      // $FlowFixMe[prop-missing]: require_tls needs to be added
       receiverConfig: receiver,
     });
   },
@@ -80,6 +81,7 @@ export const MagmaAlarmsApiUtil: ApiUtil = {
     await MagmaV1API.putNetworksByNetworkIdPrometheusAlertReceiverByReceiver({
       networkId: nullthrows(networkId),
       receiver: receiver.name,
+      // $FlowFixMe[prop-missing]: require_tls needs to be added
       receiverConfig: receiver,
     });
   },
@@ -112,6 +114,16 @@ export const MagmaAlarmsApiUtil: ApiUtil = {
       networkId: nullthrows(networkId),
     });
     return series;
+  },
+  getMetricNames: async ({networkId}) => {
+    const series = await MagmaV1API.getNetworksByNetworkIdPrometheusSeries({
+      networkId: nullthrows(networkId),
+    });
+    const names = new Set([]);
+    series.forEach(value => {
+      names.add(value.__name__);
+    });
+    return Array.from(names);
   },
 
   //alertmanager global config
