+++ nms/app/packages/magmalte/app/views/subscriber/SubscriberChart.js
@@ -15,6 +15,7 @@
  */
 import type {DataRows} from '../../components/DataGrid';
 import type {Dataset} from '../../components/CustomMetrics';
+import type {EnqueueSnackbarOptions} from 'notistack';
 import type {network_id, subscriber_id} from '@fbcnms/magma-api';
 
 import Card from '@material-ui/core/Card';
@@ -60,7 +61,10 @@ type DatasetFetchProps = {
   subscriberId: subscriber_id,
   start: moment,
   end: moment,
-  enqueueSnackbar: (msg: string, cfg: {}) => ?(string | number),
+  enqueueSnackbar: (
+    msg: string,
+    cfg: EnqueueSnackbarOptions,
+  ) => ?(string | number),
 };
 
 async function getDatasets(props: DatasetFetchProps) {
