+++ nms/app/packages/magmalte/package.json
@@ -26,7 +26,7 @@
     "@babel/register": "^7.13.16",
     "@babel/runtime": "^7.14.0",
     "@date-io/moment": "1.3.13",
-    "@fbcnms/alarms": "^0.1.12",
+    "@fbcnms/alarms": "^0.2.3",
     "@fbcnms/auth": "^0.1.0",
     "@fbcnms/babel-register": "^0.1.2",
     "@fbcnms/express-middleware": "^0.1.3",
@@ -36,7 +36,7 @@
     "@fbcnms/sequelize-models": "^0.1.8",
     "@fbcnms/strings": "^0.1.0",
     "@fbcnms/types": "^0.1.11",
-    "@fbcnms/ui": "^0.1.8",
+    "@fbcnms/ui": "^0.1.11",
     "@fbcnms/util": "^0.1.0",
     "@material-ui/core": "^4.11.4",
     "@material-ui/icons": "^4.0.0",
